@@ -1,10 +1,8 @@
 export type PathToArray = (path: string) => string[]
 
-const pathToArray: PathToArray = (path) => {
+export const pathToArray: PathToArray = (path) => {
   return path
     .split('/')
     .filter((path) => path !== '')
     .map((path) => '/' + path)
 }
-
-export default pathToArray
